// types/ui.ts
export type PPIconName =
  | 'home' | 'users' | 'groups' | 'rooms' | 'server' | 'config' | 'site' | 'plugins'
  | 'logs' | 'console' | 'audit' | 'jobs' | 'notification' | 'automation' | 'preferences'
  | 'download' | 'logout' | 'lock'
  | 'warning' | 'error' | 'live' | 'replay' | 'charts' | 'menu' | 'close' | 'chevron' | 'search'